requirements.txt
scikit-learn
rpy2==3.0.1
biopython==1.74
keras==2.2.5
numpy==1.16.5
tensorflow==1.14.0
pandas